fhc-webred/backend/config.py
#add a .env file for environment safety

import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    SECRET_KEY = os.getenv("SECRET_KEY", "your_fallback_secret")
    STRIPE_SECRET_KEY = os.getenv("STRIPE_SECRET_KEY")
    STRIPE_PUBLIC_KEY = os.getenv("STRIPE_PUBLIC_KEY")
    DEBUG = os.getenv("DEBUG", "False").lower() in ("true", "1", "t")